feat(SheetPage): show current day on the top and fetch per month
