+++ apps/web/pages/time/sheet.page.tsx
@@ -1,14 +1,5 @@
-import { startOfWeek } from 'date-fns'
-
 import { WorkHoursSheet } from '../../frontend/components/timeSheetView/workHoursSheet'
 
-const SheetPage = (): JSX.Element => {
-  const startOfTheWeek = startOfWeek(new Date(), { weekStartsOn: 1 })
+const SheetPage = (): JSX.Element => <WorkHoursSheet />
 
-  return (
-    <>
-      <WorkHoursSheet startDate={startOfTheWeek} />
-    </>
-  )
-}
 export default SheetPage
